+++ app/components/quickfile-nav/component.ts
@@ -1,0 +1,17 @@
+import Component from '@ember/component';
+
+export default class QuickfileNav extends Component.extend({
+    tagName: 'nav',
+
+    classNames: ['navbar', 'osf-project-navbar', 'row'],
+
+    attributeBindings: ['role'],
+
+    role: 'navigation',
+}) {}
+
+declare module '@ember/component' {
+    interface IRegistry {
+        'quickfile-nav': QuickfileNav;
+    }
+}