+++ app/config/environment.d.ts
@@ -1,0 +1,16 @@
+export default config;
+
+/**
+ * Type declarations for
+ *    import config from './config/environment'
+ *
+ * For now these need to be managed by the developer
+ * since different ember addons can materialize new entries.
+ */
+declare namespace config {
+  const environment: any;
+  const modulePrefix: string;
+  const podModulePrefix: string;
+  const locationType: string;
+  const rootURL: string;
+}