+++ app/locales/en/config.ts
@@ -2,6 +2,6 @@
 // can safely delete this file. Use it if you need to override behavior
 // for a locale or define behavior for a locale that Ember-I18n
 // doesn't know about.
-export default {
-    rtl: false,
-};
+export default class {
+    public rtl = false;
+}