+++ app/router.ts
@@ -1,9 +1,27 @@
 import EmberRouter from '@ember/routing/router';
-import config from './config/environment';
+import { scheduleOnce } from '@ember/runloop';
+import { inject as service } from '@ember/service';
+import config from 'ember-get-config';
 
 const Router = EmberRouter.extend({
+    metrics: service('metrics'),
+
     location: config.locationType,
     rootURL: config.rootURL,
+
+    didTransition() {
+        this._super(...arguments);
+        this._trackPage();
+    },
+
+    _trackPage() {
+        scheduleOnce('afterRender', this, () => {
+            const page = this.get('url');
+            const title = this.getWithDefault('currentRouteName', 'unknown');
+
+            this.get('metrics').trackPage({ page, title });
+        });
+    },
 });
 
 /* eslint-disable array-callback-return */