+++ config/environment.js
@@ -46,6 +46,26 @@ module.exports = function(environment) {
             },
         ],
         FB_APP_ID: process.env.FB_APP_ID,
+        microfeedback: {
+            enabled: true,
+            url: null,
+            pageParams: {
+                // Mapping of pageName to query params to add
+                // to the base MicroFeedback URL
+                // e.g. {
+                //    QuickFiles: {
+                //        componentID: '13836',
+                //        priorityID: '10100',
+                //    }
+                // }
+                QuickFiles: {},
+            },
+        },
+        social: {
+            twitter: {
+                viaHandle: 'OSFramework',
+            },
+        },
     };
 
     if (environment === 'development') {