+++ tests/unit/application/controller-test.ts
@@ -1,11 +1,14 @@
 import { moduleFor, test } from 'ember-qunit';
 
 moduleFor('controller:application', 'Unit | Controller | application', {
-    needs: ['service:session'],
+    needs: [
+        'service:session',
+        'service:metrics',
+    ],
 });
 
 // Replace this with your real tests.
-test('it exists', function(assert) {
+test('it exists', function(this: test, assert) {
     const controller = this.subject();
     assert.ok(controller);
 });